Increase column limit (#68)
* store INT96 values on the heap when boxed in a parquet.Value

* support parquet schemas with up to 32K columns

+++ buffer_test.go
@@ -238,8 +238,8 @@ func descending(typ parquet.Type, values []parquet.Value) {
 }
 
 func testBuffer(t *testing.T, node parquet.Node, reader parquet.ValueReader, buffer *parquet.Buffer, encoder encoding.Encoder, values []interface{}, sortFunc sortFunc) {
-	repetitionLevel := int8(0)
-	definitionLevel := int8(0)
+	repetitionLevel := 0
+	definitionLevel := 0
 	if !node.Required() {
 		definitionLevel = 1
 	}
